requirements.txt
discord.py
guilded.py
voltage @ git+https://github.com/EnokiUN/voltage
requests